requirements.txt

asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
channels==4.1.0
charset-normalizer==3.3.2
Django==4.2.11
djangorestframework==3.15.1
idna==3.7
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
